requirements.txt
flask
pdfplumber
rembg
playwright
playwright-stealth
pillow
pillow-avif-plugin
onnxruntime
